cppcheck: fix printf format for unsigned int bitfields
Use %u instead of %d for unsigned int bitfields kill_vg,
fence_pr and no_timeout in print_lockspace dump output.

cppcheck: cast unsigned bitfields in printf for lvmlockd

Cast unsigned bitfields to (unsigned) to match %u format
specifier, satisfying both cppcheck and gcc -Wformat-signedness.

GCC may choose unsigned underlying type for enums with
non-negative values, use %u with explicit (unsigned) cast.

+++ lib/datastruct/radix-tree-adaptive.c
@@ -1202,7 +1202,7 @@ static bool _check_nodes(struct value *v, unsigned *count)
 		return true;
 
 	default:
-		fprintf(stderr, "unknown value type: %u\n", v->type);
+		fprintf(stderr, "unknown value type: %u\n", (unsigned)v->type);
 	}
 
 	fprintf(stderr, "shouldn't get here\n");
